feat: rename layout file
