+++ src/components/_pages/PageBridge.tsx
@@ -1,5 +1,5 @@
 import { NextPage } from "next"
-import { Layout } from "components/Layout"
+import { Layout } from "components/_layout/Layout"
 import { Center } from "@chakra-ui/react"
 
 import { BridgeCard } from "components/_cards/BridgeCard"
